+++ examples/OledClock/StoredInfo.h
@@ -11,16 +11,24 @@ struct StoredInfo {
   /** 00:00:00 - 23:59:59 */
   static uint8_t const kTwentyFour = 1;
 
-  uint8_t hourMode; // either kTwelve or kTwentyFour
+  /** Either kTwelve or kTwentyFour. */
+  uint8_t hourMode;
 
-  uint8_t type; // TimeZoneData::kType*
+  /** TimeZoneData::kType*. */
+  uint8_t type;
+
   union {
+    /** Used by kTypeFixed. */
     int8_t offsetCode;
+
+    /** Used by kTypeManual. */
     struct {
       int8_t stdOffsetCode;
       bool isDst;
     };
-    uint8_t zoneIndex;
+
+    /** Used by kTypeBasic or kTypeExtended. */
+    uint16_t zoneIndex;
   };
 };
